fix bugs in nested serializable objects
inherit annotations from subclasses without modification of the
subclasses (don't add __serializable__ and don't delete default
class values).

the only time the __serializable__ cache is added, or the defaults
are removed, is the first time "fields_" is called on a class.
subclasses are inspected in reverse order in order to derive any
inherited annotations, but only the current class is modified.